+++ subgraphs/base/src/price/PriceLookup.ts
@@ -1,50 +1,18 @@
 import { Address, BigDecimal, BigInt, log } from "@graphprotocol/graph-ts";
 
 import { PriceHandler, PriceLookupResult } from "../../../shared/src/price/PriceHandler";
-import { PriceHandlerBalancer } from "../../../shared/src/price/PriceHandlerBalancer";
-import { PriceHandlerStablecoin } from "../../../shared/src/price/PriceHandlerStablecoin";
 import { PriceHandlerUniswapV2 } from "../../../shared/src/price/PriceHandlerUniswapV2";
-import { PriceHandlerUniswapV3 } from "../../../shared/src/price/PriceHandlerUniswapV3";
 import { getUSDRate } from "../../../shared/src/price/PriceRouter";
 import {
-  BALANCER_VAULT,
-  ERC20_ARB,
-  ERC20_FRAX,
-  ERC20_GOHM_SYNAPSE,
-  ERC20_JONES,
-  ERC20_LQTY,
-  ERC20_MAGIC,
   ERC20_OHM,
-  ERC20_USDC,
-  ERC20_VSTA,
   ERC20_WETH,
-  LP_BALANCER_POOL_OHM_USDC,
-  LP_BALANCER_POOL_WETH_OHM,
-  LP_BALANCER_POOL_WETH_VESTA,
-  LP_CAMELOT_OHM_WETH,
-  LP_UNISWAP_V2_GOHM_WETH,
-  LP_UNISWAP_V2_JONES_WETH,
-  LP_UNISWAP_V2_LQTY_WETH,
-  LP_UNISWAP_V2_MAGIC_WETH,
-  LP_UNISWAP_V3_ARB_WETH,
-  LP_UNISWAP_V3_WETH_USDC,
+  LP_UNISWAP_V2_OHM_WETH
 } from "../contracts/Constants";
 import { getContractName } from "../contracts/Contracts";
 import { getBaseTokenRate, isBaseToken } from "./PriceBase";
 
 export const PRICE_HANDLERS: PriceHandler[] = [
-  // new PriceHandlerBalancer([ERC20_MAGIC, ERC20_USDC], BALANCER_VAULT, LP_BALANCER_POOL_MAGIC_USDC, getContractName), // DO NOT enable: will cause infinite loop: https://github.com/OlympusDAO/olympus-protocol-metrics-subgraph/issues/94
-  new PriceHandlerBalancer([ERC20_WETH, ERC20_VSTA], BALANCER_VAULT, LP_BALANCER_POOL_WETH_VESTA, getContractName),
-  new PriceHandlerBalancer([ERC20_WETH, ERC20_OHM], BALANCER_VAULT, LP_BALANCER_POOL_WETH_OHM, getContractName),
-  new PriceHandlerBalancer([ERC20_OHM, ERC20_USDC], BALANCER_VAULT, LP_BALANCER_POOL_OHM_USDC, getContractName),
-  new PriceHandlerStablecoin([ERC20_FRAX, ERC20_USDC], getContractName),
-  new PriceHandlerUniswapV2([ERC20_GOHM_SYNAPSE, ERC20_WETH], LP_UNISWAP_V2_GOHM_WETH, getContractName),
-  new PriceHandlerUniswapV2([ERC20_JONES, ERC20_WETH], LP_UNISWAP_V2_JONES_WETH, getContractName),
-  new PriceHandlerUniswapV2([ERC20_LQTY, ERC20_WETH], LP_UNISWAP_V2_LQTY_WETH, getContractName),
-  new PriceHandlerUniswapV2([ERC20_MAGIC, ERC20_WETH], LP_UNISWAP_V2_MAGIC_WETH, getContractName),
-  new PriceHandlerUniswapV3([ERC20_USDC, ERC20_WETH], LP_UNISWAP_V3_WETH_USDC, getContractName),
-  new PriceHandlerUniswapV3([ERC20_ARB, ERC20_WETH], LP_UNISWAP_V3_ARB_WETH, getContractName),
-  new PriceHandlerUniswapV2([ERC20_OHM, ERC20_WETH], LP_CAMELOT_OHM_WETH, getContractName),
+  new PriceHandlerUniswapV2([ERC20_OHM, ERC20_WETH], LP_UNISWAP_V2_OHM_WETH, getContractName),
 ];
 
 /**