+++ subgraphs/arbitrum/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Arbitrum Subgraph Changelog
 
+## 1.7.9 (2024-07-17)
+
+- Adds POL in Camelot
+- Change the Chainlink price feed used as a trigger in order to increase indexing speed
+- Reduce code paths for determining protocol addresses
+- Modify Uniswap V2 PriceHandler to return 0 if the contract reverts
+
 ## 1.6.4 (2023-09-13)
 
 - Apply write-off of JonesDAO token, since it cannot be sold.